Reordered modules into check/modify/add/remove
