+++ demeuk.py
@@ -150,23 +150,18 @@ from html import unescape
 from inspect import cleandoc
 from locale import LC_ALL, setlocale
 from math import ceil
-from multiprocessing import cpu_count, Pool
+from multiprocess import cpu_count, Pool # multiprocess has better serialization capabilities
 from os import linesep, access, path, R_OK, F_OK, W_OK
-from re import compile as re_compile
-from re import search
-from re import split as re_split
-from re import sub
 from signal import signal, SIGINT, SIG_IGN
 from string import punctuation as string_punctuation
 from time import sleep
 from sys import stderr, stdin, stdout
-from unicodedata import category
 
 
 from chardet import detect
 from docopt import docopt
 from ftfy import fix_encoding
-from ftfy.chardata import HTML_ENTITIES, HTML_ENTITY_RE
+from ftfy.chardata import HTML_ENTITIES
 from ftfy.fixes import fix_latin_ligatures
 from nltk import str2tuple
 from nltk.tokenize import WhitespaceTokenizer
@@ -174,856 +169,18 @@ from tqdm import tqdm
 from transliterate import translit
 from unidecode import unidecode
 
+from modules.modify import *
+from modules.check import *
+from modules.remove import *
+from modules.add import *
+
 
 version = '4.6.2'
 
-# Search from start to finish for the string $HEX[], with block of a-f0-9 with even number
-# of hex chars. The first match group is repeated.
-HEX_REGEX = re_compile(r"^\$(?:HEX|hex)\[((?:[0-9a-fA-F]{2})+)\]$")
-EMAIL_REGEX = '.{1,64}@([a-zA-Z0-9_-]{1,63}\\.){1,3}[a-zA-Z]{2,6}'
-HASH_HEX_REGEX = '^[a-fA-F0-9]+$'
-MAC_REGEX = '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
-UUID_REGEX = '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
 
-# Officiale bcrypt hashes hae a bit more fixed size, but saw some weird once:
-# $2a$10$demo as example
-HASH_BCRYPT_REGEX = '^\\$2[ayb]\\$[0-9]{1,}\\$[\\w\\.\\/]{4,}$'
-# Crypt hashes can look a lot like passwords. We do two options here
-# $1[$optional salt, max 16]$string of a-zA-Z0-9./ length 7 min till end of line
-# $1$a-zA-Z0-9./ min length 12 to make sure we hit somthing like: a-zA-Z0-9./
-# this will cause string like $1$JAjdna./d to still be included.
 
-HASH_CRYPT_REGEX = '^\\$[1356]\\$[\\w\\.\\/]{12,}$'
-HASH_CRYPT_SALT_REGEX = '^\\$[1356]\\$[\\w\\.\\/\\+]{,16}\\$[\\w\\.\\/]{6,}$'
-HASH_PHPBB_REGEX = '^\\$[hH]\\$[\\w\\.\\/]{6,}$'
-HASH_REGEX_LIST = [HASH_BCRYPT_REGEX, HASH_CRYPT_SALT_REGEX, HASH_CRYPT_REGEX, HASH_PHPBB_REGEX]
-
-TRIM_BLOCKS = ('\\\\n', '\\\\r', '\\n', '\\r', '<br>', '<br />')
 
 CHUNK_SIZE = 1024 * 1024
-
-
-def _unescape_fixup_named(match):
-    """
-    Replace one matched HTML entity with the character it represents,
-    if possible.
-
-    Based on: ftfy.fixes._unescape_fixup
-    """
-    text = match.group(0)
-    if text in HTML_ENTITIES:
-        return HTML_ENTITIES[text]
-    else:
-        return text
-
-
-def _unescape_fixup(match):
-    """
-    Replace one matched HTML entity with the character it represents,
-    if possible.
-
-    Based on: ftfy.fixes._unescape_fixup
-    """
-    text = match.group(0)
-    if text.startswith('&#'):
-        unescaped = unescape(text)
-
-        # If html.unescape only decoded part of the string, that's not what
-        # we want. The semicolon should be consumed.
-        if ';' in unescaped:
-            return text
-        else:
-            return unescaped
-    else:
-        return text
-
-
-def clean_googlengram(line):
-    """Removes speechtags from line specific to the googlengram module
-
-    Param:
-        line (unicode)
-
-    Returns:
-        line (unicode)
-    """
-    return_line = line.split("\t")[0]  # Get the ngram, remove year, counter, etc
-    clean = []
-    words = WhitespaceTokenizer().tokenize(return_line)
-    for word in words:
-        # in >1-grams transitions to specific tags are written as:
-        # The_ADJ _NOUN_ (meaning from The there is a transition to a noun
-        # We remove those
-        if word[0] != '_' and word[-1] != '_':
-            # Split the token and the tag based on the '_'
-            token, tag = str2tuple(word, '_')
-            # Punct will be added using rules.
-            if len(token) > 1:
-                if tag != 'PUNCT' or tag != '.' or tag != '':
-                    clean.append(token)
-            elif token not in string_punctuation:
-                clean.append(token)
-    return_line = ' '.join(clean)
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def remove_email(line):
-    """Removes e-mail addresses from a line.
-
-    Params:
-        line (unicode)
-
-    Returns:
-        line (unicode)
-    """
-    if '@' in line:
-        if search(f'{EMAIL_REGEX}(:|;)', line):
-            return True, sub(f'{EMAIL_REGEX}(:|;)', '', line)
-    return False, line
-
-
-def add_lower(line):
-    """Returns if the upper case string is different from the lower case line
-
-    Param:
-        line (unicode)
-
-    Returns:
-        False if they are the same
-        Lowered string if they are not
-    """
-    line_lower = line.lower()
-    if line != line_lower:
-        return line_lower
-    else:
-        return False
-
-
-def add_first_upper(line):
-    """Returns the line with the first letter capitalized and all the others in lowercase.
-
-    Param:
-        line (unicode)
-
-    Returns:
-        False if they are the same
-        Capitalized string if they are not
-    """
-    line_first_upper = line.capitalize()
-    if line != line_first_upper:
-        return line_first_upper
-    else:
-        return False
-
-
-def add_title_case(line):
-    """Returns title case string where all the first letters are capitals and all others in lowercase.
-
-    Param:
-        line (unicode)
-
-    Returns:
-        False if they are the same
-        Title string if they are not
-    """
-    line_title_case = line.title()
-    if line != line_title_case:
-        return line_title_case
-    else:
-        return False
-
-
-def add_latin_ligatures(line):
-    """Returns the line cleaned of latin ligatures if there are any.
-
-    Param:
-        line (unicode)
-
-    Returns:
-        False if there are not any latin ligatures
-        Corrected line
-    """
-    cleaned_line = fix_latin_ligatures(line)
-    if line != cleaned_line:
-        return cleaned_line
-    else:
-        return False
-
-
-def add_without_punctuation(line, punctuation):
-    """Returns the line cleaned of punctuation.
-
-    Param:
-        line (unicode)
-
-    Returns:
-        False if there are not any punctuation
-        Corrected line
-    """
-    cleaned_line = line.translate(str.maketrans('', '', punctuation))
-
-    if line != cleaned_line:
-        return cleaned_line
-    else:
-        return False
-
-
-def clean_add_umlaut(line):
-    """Returns the line cleaned of incorrect umlauting
-
-    Param:
-        line (unicode)
-
-    Returns:
-        Corrected line
-    """
-    cleaned_line = line
-
-    umlaut_dict = {
-        'a"': 'ä',
-        'i"': 'ï',
-        'o"': 'ö',
-        'u"': 'ü',
-        'e"': 'ë',
-        'A"': 'Ä',
-        'I"': 'Ï',
-        'O"': 'Ö',
-        'U"': 'Ü',
-        'E"': 'Ë',
-    }
-    for letter in umlaut_dict.keys():
-        cleaned_line = cleaned_line.replace(letter, umlaut_dict.get(letter))
-
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def remove_punctuation(line, punctuation):
-    """Returns the line without punctuation
-
-    Param:
-        line (unicode)
-        punctuation (unicode)
-
-    Returns:
-        line without start and end punctuation
-    """
-    return_line = line.translate(str.maketrans('', '', punctuation))
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def remove_strip_punctuation(line, punctuation):
-    """Returns the line without start and end punctuation
-
-    Param:
-        line (unicode)
-
-    Returns:
-        line without start and end punctuation
-    """
-    return_line = line.strip(punctuation)
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def add_split(line, punctuation=(' ', '-', r'\.')):
-    """Split the line on the punctuation and return elements longer then 1 char.
-
-    Param:
-        line (unicode)
-
-    Returns:
-        split line
-    """
-    for p in punctuation:
-        if p in line:
-            return [i for i in re_split('|'.join(punctuation), line) if len(i) > 1]
-    return False
-
-
-def check_case(line, ignored_chars=(' ', "'", '-')):
-    """Checks if an uppercase line is equal to a lowercase line.
-
-    Param:
-        line (unicode)
-        ignored_chars list(string)
-
-    Returns:
-        true if uppercase line is equal to uppercase line
-    """
-    for c in line:
-        c = str(c)
-        if c.lower() == c.upper():
-            if c in ignored_chars:
-                continue
-            else:
-                return False, c
-    return True, None
-
-
-def check_length(line, min=0, max=0):
-    """Does a length check on the line
-
-    Params:
-        line (unicode)
-        min (int)
-        max (int)
-
-    Returns:
-        true if length is ok
-    """
-    status = True
-    if min and status:
-        status = len(line) >= min
-    if max and status:
-        status = len(line) < max
-    return status
-
-
-def check_hash(line):
-    """Check if a line contains a hash
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true if line does not contain hash
-    """
-    if search(HASH_HEX_REGEX, line):
-        if len(line) in [32, 40, 64]:
-            return False
-    if len(line) > 0:
-        if line[0] == '$':
-            for hash_regex in HASH_REGEX_LIST:
-                if search(hash_regex, line):
-                    return False
-    return True
-
-
-def check_mac_address(line):
-    """Check if a line contains a MAC-address
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true if line does not contain a MAC-address
-    """
-    if search(MAC_REGEX, line):
-        return False
-
-    return True
-
-
-def check_email(line):
-    """Check if lines contain e-mail addresses with a simple regex
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true is line does not contain email
-    """
-    if search(EMAIL_REGEX, line):
-        return False
-    else:
-        return True
-
-
-def check_non_ascii(line):
-    """Checks if a line contains a non ascii chars
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true if line does not contain non ascii chars
-    """
-    try:
-        line.encode('ascii')
-        return True
-    except UnicodeEncodeError:
-        return False
-
-
-def check_character(line, character):
-    """Checks if a line contains a specific character
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true if line does contain the specific character
-
-    """
-    if character in line:
-        return True
-    else:
-        return False
-
-
-def check_starting_with(line, strings):
-    """Checks if a line start with a specific strings
-
-    Params:
-        line (unicode)
-        strings[str]
-
-    Returns:
-        true if line does start with one of the strings
-
-    """
-    for string in strings:
-        if line.startswith(string):
-            return True
-    return False
-
-
-def check_uuid(line):
-    """Check if a line contains a UUID
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true if line does not contain a UUID
-    """
-    if search(UUID_REGEX, line):
-        return False
-
-    return True
-
-
-def check_ending_with(line, strings):
-    """Checks if a line ends with specific strings
-
-    Params:
-        line (unicode)
-        strings[str]
-
-    Returns:
-        true if line does end with one of the strings
-
-    """
-    for string in strings:
-        if line.endswith(string):
-            return True
-    return False
-
-
-def check_contains(line, strings):
-    """Checks if a line does not contain specific strings
-
-    Params:
-        line (unicode)
-        strings[str]
-
-    Returns:
-        true if line does contain any one of the strings
-
-    """
-    for string in strings:
-        if string in line:
-            return True
-    return False
-
-
-def check_empty_line(line):
-    """Checks if a line is empty or only contains whitespace chars
-
-    Params:
-        line (unicode)
-
-    Returns:
-        true of line is empty or only contains whitespace chars
-    """
-    if line == '':
-        return True
-    elif line.isspace():
-        return True
-    return False
-
-
-def clean_cut(line, delimiters, fields):
-    """Finds the first delimiter and returns the remaining string either after
-    or before the delimiter.
-
-    Params:
-        line (unicode)
-        delimiters list(unicode)
-        fields (unicode)
-
-    Returns:
-        line (unicode)
-    """
-    for delimiter in delimiters:
-        if delimiter in line:
-            if '-' in fields:
-                start = fields.split('-')[0]
-                stop = fields.split('-')[1]
-                if start == '':
-                    start = 1
-                if stop == '':
-                    stop = len(line)
-                fields = slice(int(start) - 1, int(stop))
-            else:
-                fields = slice(int(fields) - 1, int(fields))
-            return True, delimiter.join(line.split(delimiter)[fields])
-    else:
-        return False, line
-
-
-def clean_transliterate(line, language):
-    """Transliterate a string
-
-    Params:
-        line (Unicode)
-        language (str)
-
-    Returns:
-        line (Unicode)
-    """
-    cleaned_line = translit(line, language, reversed=True)
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def clean_non_ascii(line):
-    """Replace non ascii chars with there ascii representation.
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-    """
-    cleaned_line = unidecode(line)
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def clean_lowercase(line):
-    """Replace all capitals to lowercase
-
-        Params:
-            line (Unicode)
-
-        Returns:
-            line (Unicode)
-
-        """
-    cleaned_line = line.lower()
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def clean_title_case(line):
-    """Replace words to title word (uppercasing first letter)
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-
-    """
-    cleaned_line = line.title()
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def clean_trim(line):
-    """Delete leading and trailing character sequences representing a newline
-    from beginning end end of line.
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-    """
-    cleaned_line = line
-    # Ensure removal of duplicated blocks
-    while True:
-        has_match = False
-        for x in TRIM_BLOCKS:
-            if cleaned_line.startswith(x):
-                cleaned_line = cleaned_line[len(x):]
-                has_match = True
-
-            if cleaned_line.endswith(x):
-                cleaned_line = cleaned_line[:-len(x)]
-                has_match = True
-
-        if has_match is False:
-            break
-
-    if line != cleaned_line:
-        return True, cleaned_line
-    else:
-        return False, line
-
-
-def clean_tab(line):
-    """Replace tab character with ':' greedy
-
-    Params:
-        line (bytes)
-
-    Returns:
-        line (bytes)
-    """
-    if b'\x09' in line:
-        line = sub(b'\x09+', b'\x3a', line)
-        return True, line
-    else:
-        return False, line
-
-
-def clean_hex(line):
-    """Converts strings like '$HEX[]' to proper binary
-
-    Params:
-        line (bytes)
-
-    Returns:
-        line (bytes)
-    """
-    match = HEX_REGEX.search(line)
-    if match:
-        return True, unhexlify(match.group(1))
-    else:
-        return False, line
-
-
-def clean_html(line):
-    """Detects html encode chars and decodes them
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-    """
-    return_line = HTML_ENTITY_RE.sub(_unescape_fixup, line)
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def clean_html_named(line):
-    """Detects named html encode chars and decodes them
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-    """
-    return_line = HTML_ENTITY_RE.sub(_unescape_fixup_named, line)
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def clean_newline(line):
-    """Delete newline characters at start and end of line
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        line (Unicode)
-    """
-    return_line = line.strip('\r\n')
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def check_controlchar(line):
-    """Detects control chars, returns True when detected
-
-    Params:
-        line (Unicode)
-
-    Returns:
-        Status, String
-    """
-    for c in line:
-        # https://en.wikipedia.org/wiki/Unicode_character_property#General_Category
-        # Characters (they have meaning):
-        # Cc -> Control Char (End of stream)
-        # Cf -> Control flow (right to left)
-        # Non chars:
-        # Cn -> Not assigned
-        # Co -> Private use
-        # Cs -> Surrogate
-        if category(c) in ['Cc', 'Cf', 'Cn', 'Co', 'Cs']:
-            return True, c
-    return False, None
-
-
-def check_regex(line, regex):
-    """Checks if a line matches a list of regexes
-
-    Params:
-        line (unicode)
-        regex (list)
-
-    Returns:
-        true if all regexes match
-        false if line does not match regex
-    """
-    for regex in regex:
-        if search(regex, line):
-            continue
-        else:
-            return False
-    return True
-
-
-def contains_at_least(line, bound, char_property):
-    """Check if the line contains at least `bound` characters with given property.
-
-    Params:
-        line (unicode)
-        bound (int)
-        char_property (str -> bool)
-
-    Returns:
-        true if at least `bound` characters match
-        false otherwise
-    """
-    if bound == 0:
-        return True
-
-    count = 0
-    for char in line:
-        if char_property(char):
-            count += 1
-            if count >= bound:
-                return True
-    return False
-
-
-def contains_at_most(line, bound, char_property):
-    """Check if the line contains at most `bound` characters with given property.
-
-    Params:
-        line (unicode)
-        bound (int)
-        char_property (str -> bool)
-
-    Returns:
-        true if at most `bound` characters match
-        false otherwise
-    """
-    count = 0
-    for char in line:
-        if char_property(char):
-            count += 1
-            if count > bound:
-                return False
-    return True
-
-
-def try_encoding(line, encoding):
-    """Tries to decode a line using supplied encoding
-
-    Params:
-        line (Byte): byte variable that will be decoded
-        encoding (string): the encoding to be tried
-
-    Returns:
-        False if decoding failed
-        String if decoding worked
-    """
-    try:
-        # Try to decode the line
-        line_decoded = line.decode(encoding)
-        # Some encoding will decoded almost any line, lets check if we have invalid chars.
-        # If we have invalid chars (except for tab like chars) we will fail
-        for c in line_decoded:
-            if category(c) in ['Cc', 'Cf', 'Cn', 'Co', 'Cs']:
-                if c == '\t' or c == '\f':
-                    continue
-                else:
-                    return False
-        return line_decoded
-    except UnicodeDecodeError:
-        return False
-
-
-def clean_mojibake(line):
-    """Detects mojibake and tries to correct it.
-    Mojibake are string that are decoded incorrectly and then encoded incorrectly.
-    This results in strings like: Ãºnico which should be único.
-
-    Param:
-        line (str)
-
-    Returns:
-        Cleaned string
-    """
-    return_line = fix_encoding(line)
-    if return_line != line:
-        return True, return_line
-    else:
-        return False, line
-
-
-def clean_encode(line, input_encoding):
-    """Detects and tries encoding
-
-    Params:
-        line (bytes)
-
-    Returns:
-        Decoded UTF-8 string
-    """
-    # Try either a user set of encodings or the default encoding set.
-    # When using multiple encoding is it beter to have multibyte encodings before
-    # Single byte encodings. Also it is beter to not include iso encoding by default.
-    # https://en.wikipedia.org/wiki/Character_encoding#Common_character_encodings
-    # Input_encoding is by default [utf8]
-    for encoding in input_encoding:
-        line_decoded = try_encoding(line, encoding)
-        if line_decoded is not False:
-            break
-    # All other methods failed, lets run the detect library on the line and try to guess the encoding.
-    if line_decoded is False:
-        encode = detect(line)
-        if encode.get('encoding'):
-            try:
-                line_decoded = line.decode(encode['encoding'])
-            except (UnicodeDecodeError, LookupError) as e: # noqa F841
-                return False, encode["encoding"]
-        else:
-            return False, 'Unknown'
-    # If we managed to get here, return decode line
-    return True, line_decoded
 
 
 def clean_up(lines):
